Add live server/worlds overview using cgi
For this to be enabled, executable .cgi files have to be passed to
fcgiwrap by nginx.
Using the 'mcstatus' python library by Dinnerbone, the servers are
queried once per request and their status is written to the resulting
html page using the mustache template engine (using the chevron package
on python).